pyproject.toml
[tool.poetry]
name = "repl_python3_RSA"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
sympy = "^1.8"
factordb-pycli = "^1.3.0"
rsa = "^4.9"
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
